Centralized wam methods into one module, introduced potential wam stencil class
